+++ src/main/scala/ai/starlake/job/ingest/AuditLog.scala
@@ -71,6 +71,7 @@ case class AuditLog(
 ) {
   def asSelect(engineName: Engine)(implicit settings: Settings): String = {
     import ai.starlake.utils.Formatter._
+    timestamp.setNanos(0)
     val template = settings.appConfig.jdbcEngines
       .get(engineName.toString.toLowerCase())
       .flatMap(_.tables("audit").selectSql)